% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoDE.R
\name{autoDE}
\alias{autoDE}
\title{Automatic Differential Expression Analysis using DESeq2}
\usage{
autoDE(sampleTable = NULL, countTable = NULL, colData = NULL, retExplore = F)
}
\arguments{
\item{sampleTable}{Filename indicating the sampleTable tab delimited file pointing to HTSeq-count files.}

\item{countTable}{Filename indicating the raw count table tab delimited file (use for featureCounts output and with colData)}

\item{colData}{Filename indicating the tab delimited file containing metadata for countTable}

\item{retExplore}{Boolean indicating if BinfTools::exploreData() function should also be returned. Requires gene symbols, and BinfTools/gpGeneSets packages. Default=F}
}
\value{
List with DESeq2 analyzed data: results, normalized counts, and conditions
}
\description{
Obtain DESeq2 Normalized results
}
