% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoDE.R
\name{autoDE}
\alias{autoDE}
\title{Automatic Differential Expression Analysis using DESeq2}
\usage{
autoDE(sampleTable = NULL, countTable = NULL, colData = NULL, expFilt=0, retExplore = F, retDDS=F, detID=T)
}
\arguments{
\item{sampleTable}{Filename indicating the sampleTable tab delimited file pointing to HTSeq-count files.}

\item{countTable}{Filename indicating the raw count table tab delimited file (use for featureCounts output and with colData)}

\item{colData}{Filename indicating the tab delimited file containing metadata for countTable}

\item{expFilt}{Numeric indicating the minimum average gene counts to consider a gene as being 'expressed'. If whole number, average gene count across all samples must be greater than this number. Default=0.}

\item{retExplore}{Boolean indicating if BinfTools::exploreData() function should also be returned. Requires gene symbols, and BinfTools/gpGeneSets packages. Default=F}

\item{retDDS}{Boolean indicating if the DESeq2 dataset (dds) should be returned for further usage. Defaults to FALSE.}

\item{detID}{Boolean indicating if ENSEMBL or Flybase gene IDs should be automatically detected for conversion to symbos using BinfTools::getSym(). Defatul=T}
}
\value{
List with DESeq2 analyzed data: results, normalized counts, and conditions
}
\description{
Obtain DESeq2 Normalized results
}
